ao-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ao-core"
version = "0.1.0"
description = "Agent Orchestration SDK for DSAI applications"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "openai>=1.0",
    "langfuse>=2.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "redis>=5.0",
    "psycopg[binary]>=3.1",
    "pgvector>=0.3",
    "azure-identity>=1.15",
    "azure-servicebus>=7.11",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["ao"]
